Added task tracking distance

+++ app/src/main/java/com/vps/android/interactors/task/request/StopTaskRequest.kt
@@ -2,5 +2,5 @@ package com.vps.android.interactors.task.request
 
 data class StopTaskRequest(
     val end_time: String, // "2021-10-06 16:18:57"
-    val distance: Int? = null, // Distance in km
+    val distance: Double? = null, // Distance in km
 )
